pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ciallo-agent"
version = "0.1.0"
description = "LLM-assisted natural language to Cello v2 workflow scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "jsonschema>=4.23.0",
  "openai>=1.0.0",
  "pydantic>=2.8.0",
  "pypdf>=6.0.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ciallo_agent = ["web_assets/*"]
